update index.html
I have added a CalculateARC function which is currently the same as the CalculateLOS function but with renamed global variables. It for some reason doesn't work and ends up saying that any pair of coordinates are out of range from one another, and provides only values of NaN as it steps through the map. Can you take a look and see if you can find out why this function isn't working. Another thing I added was the code which tells the user which direction the attack is coming from. Right now it only works outputting radians, but once I fix the switch statement (can you also take a look at that? It's a switch statement with the parameter 'endRadians'), it will tell us which hexagon "face"/side the attack is coming from. Also since only the LOS function is working right now, that is where the attack direction code is located.